Flytta classer till v02_models för bättre överblick
